requirements.txt
beautifulsoup4
python-dateutil
selenium
webdriver-manager
google-auth
google-auth-oauthlib
google-api-python-client
